Move populate script tests
